[#1087] Split logic of ProgressBar component into separate components.
