+++ frontend/src/common/progress-bar/Spinner.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react';
+import * as classNames from 'classnames';
+
+interface SpinnerProps {
+  isVisible: boolean;
+}
+
+const Spinner: React.StatelessComponent<SpinnerProps> = ({ isVisible }) => {
+  const spinnerClassName = classNames({
+    'ProgressBar-spinner': true,
+    'hide': !isVisible,
+  });
+
+  return (
+    <div className={spinnerClassName}></div>
+  );
+};
+
+export default Spinner;
