+++ frontend/src/common/progress-bar/SpinnerIcon.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+const SpinnerIcon: React.StatelessComponent<{}> = () => (
+  <div className='ProgressBar-spinner-icon'></div>
+);
+
+export default SpinnerIcon;
